feat(dcb): remove DcbDomainTypes from TagConsistentGrain and TagStateGrain
Complete the WASM-friendly grain boundary by removing DcbDomainTypes
references from the remaining tag grains. DcbDomainTypes is now
decomposed into ITagTypes, ITagProjectorTypes, and ITagStatePayloadTypes
interfaces, keeping all domain/serializer concerns inside the
IProjectionActorHost boundary.

Closes #913
